The User-Friendliness Update
This commit has genjob being updated for ease of use and portability. There are a number of variables at the top of the file that can be set to change the input files, output files, and replay directory used. The README has been updated to reflect these changes and describe the use of these variables.

genjob should now be easily modifiable to meet the needs of most experiments.